@@ -1,36 +1,35 @@
 {
-    "name": "water-ui",
-    "version": "0.1.7",
+    "name": "water-next",
+    "version": "0.0.1",
     "private": false,
-    "main": "lib/water-ui.common.js",
     "description": "A Component Library for Vue.js",
-    "license": "MIT",
     "author": "1548319928@qq.com",
     "scripts": {
+        "serve": "vue-cli-service serve",
         "build": "vue-cli-service build",
+        "lint": "vue-cli-service lint",
         "docs:build": "vuepress build docs",
         "docs:dev": "vuepress dev docs",
-        "lib": "vue-cli-service build --target lib --name water-ui --dest lib ./components/index.js",
-        "lint": "vue-cli-service lint",
-        "serve": "vue-cli-service serve"
+        "lib": "vue-cli-service build --target lib --name water-next --dest lib ./components/index.js"
     },
+    "main": "components/water-ui.common.js",
     "dependencies": {
+        "@vue/babel-plugin-jsx": "^1.0.0-rc.2",
         "core-js": "^3.6.5",
-        "vue": "^2.6.11",
-        "vue-router": "^3.2.0",
-        "vuex": "^3.4.0"
+        "vue": "^3.0.0-beta.1"
     },
     "devDependencies": {
         "@vue/cli-plugin-babel": "~4.4.0",
         "@vue/cli-plugin-eslint": "~4.4.0",
         "@vue/cli-service": "~4.4.0",
+        "@vue/compiler-sfc": "^3.0.0-beta.1",
         "babel-eslint": "^10.1.0",
         "eslint": "^6.7.2",
-        "eslint-plugin-vue": "^6.2.2",
+        "eslint-plugin-vue": "^7.0.0-alpha.0",
         "less-loader": "^6.2.0",
         "node-sass": "^4.12.0",
         "sass-loader": "^8.0.2",
-        "vue-template-compiler": "^2.6.11",
+        "vue-cli-plugin-vue-next": "~0.1.4",
         "vuepress": "^1.5.2"
     },
     "eslintConfig": {
@@ -39,7 +38,7 @@
             "node": true
         },
         "extends": [
-            "plugin:vue/essential",
+            "plugin:vue/vue3-essential",
             "eslint:recommended"
         ],
         "parserOptions": {
@@ -51,5 +50,6 @@
         "> 1%",
         "last 2 versions",
         "not dead"
-    ]
+    ],
+    "license": "MIT"
 }
