@@ -1,17 +1,17 @@
 <template>
-  <wr-overlay v-model="show" :mask="mask" :zIndex="zIndex" :custom-style="getOverlayStyle">
+  <wr-overlay v-model:value="show" :mask="mask" :zIndex="zIndex" :custom-style="getOverlayStyle">
     <!-- <div
       class="wr-action-sheet__warp"
       v-show="show"
       @click="hide()"
       :style="getactionSheetWarpStyle"
     >-->
-    <wr-transition-animation name="bottom">
+    <transition name="bottom">
       <div
         class="wr-action-sheet"
         :style="getactionSheetStyle"
-        @click.stop=""
-  
+        @click.stop
+        @touchstart.stop
         v-show="show"
       >
         <div class="wr-action-sheet_item wr-action-sheet-description" v-if="description">
@@ -21,11 +21,7 @@
           class="wr-action-sheet_item"
           v-for="(item,index) in list"
           :key="index"
-          name="sheetItem"
-          :index="index"
-          :class="{'wr-action-sheet_item--disabled':item.disabled,'wr-action-sheet_item--active':active==index}"
-          @touchstart.stop.prevent="ItemTouchstart(index)"
-          @touchend="ItemTouchend($event,item,index)"
+          :class="{'wr-action-sheet_item--disabled':item.disabled}"
           @click="actionSheetItemClick(item)"
         >
           <div class="wr-action-sheet_item-text" :style="setActionSheetItemStyle(item)">
@@ -40,19 +36,22 @@
           </div>
         </template>
       </div>
-    </wr-transition-animation>
+    </transition>
   </wr-overlay>
   <!-- </div> -->
 </template>
 <script>
+import { computed, reactive, ref, watch } from "vue";
+import events from "../../../utils/event";
 import wrOverlay from "../../overlay/index";
-import wrTransitionAnimation from "../../transition-animation/index";
+let that;
+
 export default {
   name: "wrActionSheet",
   components: {
     wrOverlay,
-    wrTransitionAnimation,
   },
+
   // props: {
   //   //显示与隐藏
   //   show: {
@@ -86,98 +85,119 @@ export default {
   //     default: "取消"
   //   }
   // },
-  data() {
-    return {
-      show: false,
-      zIndex: 1002,
-      description: "",
-      //菜单列表
-      list: [],
-      //上边角弧度
-      borderTopRadius: 12,
-      hideCancel: false,
-      cancelText: "取消",
-      //点击遮罩后不关闭
-      mask: false,
-      //点击具体项后不关闭
-      closeOnItem: false,
-      hideOverlay: false,
-      active: -99,
-    };
+  created() {
+    that = this;
   },
-  computed: {
-    getOverlayStyle() {
-      let style = {};
-      if (this.hideOverlay) {
-        style.background = "transparent";
-      }
-      return style;
-    },
-    getactionSheetWarpStyle() {
-      return {
-        zIndex: this.zIndex,
-      };
-    },
-    getactionSheetStyle() {
-      return {
-        borderTopLeftRadius: this.borderTopRadius + "px",
-        borderTopRightRadius: this.borderTopRadius + "px",
-      };
-    },
-  },
-  watch: {
-    show: {
-      handler(val) {
-        this.$emit("change", val);
-        if (val) {
-          this.$emit("open", val);
-        } else {
-          this.$emit("close", val);
-        }
-      },
-      deep: true,
-    },
-  },
-  methods: {
-    setDate(data) {
+  setup(props, context) {
+
+    //显示
+    const show = ref(false);
+    //层级
+    const zIndex = ref(1002);
+    //头部描述
+    const description = ref("");
+    //顶部边框弧度大小
+    const borderTopRadius = ref(12);
+    //隐藏取消按钮
+    const hideCancel = ref(false);
+    //取消按钮文字
+    const cancelText = ref("取消");
+    //点击遮罩后不关闭
+    const mask = ref(false);
+    //点击具体项后不关闭
+    const closeOnItem = ref(false);
+    //隐藏遮罩
+    const hideOverlay = ref(false);
+    //选项列表
+    const list = reactive([]);
+    //事件中心
+    const event = reactive(events);
+
+    const setDate = (data) => {
       for (let key in data) {
-        key in this.$data ? (this[key] = data[key]) : "";
+        key in that ? (that[key] = data[key]) : "";
       }
-    },
-    hide() {
-      if (this.mask) return;
-      this.show = false;
-    },
-    ItemTouchstart(index) {
-      this.active = index;
-    },
-    ItemTouchend(e, item, index) {
-      this.active = -99;
-      let x = e.changedTouches[0].pageX;
-      let y = e.changedTouches[0].pageY;
-      let element = document.elementFromPoint(x, y);
-      let pName = element.getAttribute("name");
-      let pIndex = element.getAttribute("index");
-      if (pName == "sheetItem" && pIndex == index) {
-        this.show = false;
-        this.$emit("select", item);
-      }
-    },
-    actionSheetItemClick(item) {
-      if (this.closeOnItem) return; 
-      this.active = -99;
-      this.show = false;
-      this.$emit("select", item);
-    },
-    cancelClick() {
-      this.show = false;
-      this.$emit("cancel");
-    },
-    setActionSheetItemStyle(item) {
+    };
+    const hide = () => {
+      if (mask.value) return;
+      show.value = false;
+    };
+    const actionSheetItemClick = (item) => {
+      if (closeOnItem.value) return;
+      show.value = false;
+      context.emit("select", item);
+      event.emit("select", item);
+    };
+    const cancelClick = () => {
+      show.value = false;
+      context.emit("cancel");
+      event.emit("cancel");
+    };
+    const setActionSheetItemStyle = (item) => {
       return {
         color: item.color,
       };
-    },
+    };
+
+    const getOverlayStyle = computed(() => {
+      let style = {};
+      if (hideOverlay.value) {
+        style.background = "transparent";
+      }
+      return style;
+    });
+
+    const getactionSheetWarpStyle = computed(() => {
+      return {
+        zIndex: zIndex.value,
+      };
+    });
+
+    const getactionSheetStyle = computed(() => {
+      return {
+        borderTopLeftRadius: borderTopRadius.value + "px",
+        borderTopRightRadius: borderTopRadius.value + "px",
+      };
+    });
+
+    watch(
+      () => {
+        return show.value;
+      },
+      (val) => {
+        context.emit("change", val);
+        event.emit("change", val);
+        if (val) {
+          context.emit("open", val);
+          event.emit("open", val);
+        } else {
+          event.emit("close", val);
+          context.emit("close", val);
+        }
+      }
+    );
+
+    return {
+      show,
+      zIndex,
+      description,
+      borderTopRadius,
+      hideCancel,
+      cancelText,
+      mask,
+      closeOnItem,
+      hideOverlay,
+      list,
+      event,
+      setDate,
+      hide,
+      actionSheetItemClick,
+      cancelClick,
+      setActionSheetItemStyle,
+      getOverlayStyle,
+      getactionSheetWarpStyle,
+      getactionSheetStyle,
+    };
   },
 };
 </script>
