@@ -1,13 +1,6 @@
-import Vue from 'vue'
+import { createApp } from 'vue';
 import App from './App.vue'
-import router from './router'
-import store from './store'
 import waterUi from './components/water-ui/index'
-Vue.config.productionTip = false
-
-Vue.use(waterUi)
-new Vue({
-    router,
-    store,
-    render: h => h(App)
-}).$mount('#app')
+const app = createApp(App)
+app.use(waterUi)
+app.mount('#app')
