@@ -1,4 +1,4 @@
-import WrCircle from './src/circle.vue'
+import WrCircle from './src/circle.jsx'
 WrCircle.install = function(Vue) {
     Vue.component(WrCircle.name, WrCircle)
 }
